Copy config/*.ini files to MacOSX bundle.
Now every .ini file on config folder will be automatically copied to the
bundle. Please note that those files are only identified on cmake
config-time, thus the build system will not automatically detect if a
new file is added to that folder. Therefore cmake will have to be rerun
every time a new file is added.