sslworld: project airbone balls, for #131
This commit projects airbone balls the same way as they would be projected in ssl-vision.
It also disables writing set_z on the SSL_DetectionBall* to mirror ssl-vision even further.

As #131 suggests to make this feature optional, the old behavior will be used if
PROJECT_AIRBONE_BALL is set to false and grSim is recompiled.

I have no experience with grSims UI-based configurations, so I omitted adding it as an option
in the UI.