Shortcut to exit grSim and different robot carcase color for each team
Now the robots of yellow and blue team can have different "uniform" colors changing the values of `ROBOT_BLUE_CHASSIS_COLOR` and `ROBOT_YELLOW_CHASSIS_COLOR`.

The color will afffect on the texture of the robot

I still pretend to create a more conveniente way to change it the colors.