build protobuf as ExternalProject if version is too new
Since at least Protobuf 3.21 the version is incompatible causing various
include and linker errors. Maybe the incompatibility was introduced before that,
but I don't know. If that is the case the version number can just be set lower
than that in the future.
The approach is pretty much just copied from ER-Forces framework
(https://github.com/robotics-erlangen/framework) and adapted to fit the
structure of this repository.