fix github pipeline for macos and windows
Also changed vcpkg dependency from qt5 to qt5-base, because that should be
enough and vcpkg seems to build from source (?) and no one has time to build Qt
completely.
Includes some hacks to get around the fact that the macos pipeline does not set
the path to the QtConfig.cmake correctly.