a few steps towards a working windows pipeline
Windows pipeline is now doing what the samples from run-vcpkg and run-cmake
suggest it should do and vcpkg claims that it should handle cmake find_package
calls correctly if you use their vcpkg.cmake as the toolchain.
It's not working, because for some reason the find_package call for Qt still
fails with cmake claiming it can't find a Qt5config.cmake file.