fix ninja build for all ExternalProjects
Apparently some weird issues are caused by using the INSTALL_DIR parameter
of ExternalProjects when using the Ninja generator.
Makes you wonder why that is even there, but cmake works in mysterious ways,
I guess.
Adding the install part as a substep and explicitly telling cmake that the
necessary libs and executables are produced by this step fixes this issue.